ci: fix packaging of release packages
The 25b8cf2f082c0c7719b454abbd0e9fdf241bc194 changes to the build commands
resulted in the package repackaged for vst-native-modules-release did not
conatain the wasm files. This is bug in the conan recipe, but touching the
recipe at this point is not an option as that breaks the repo, so until things
have settled around the toolchain update, just revert to the old build
commands.